+++ js-miniapp-sdk/src/index.ts
@@ -9,7 +9,12 @@ import { AdTypes } from './types/adTypes';
 import { InterstitialAdResponse } from './types/responseTypes/interstitial';
 import { RewardedAdResponse } from './types/responseTypes/rewarded';
 import { ShareInfoType } from './types/ShareInfoType';
-import { CustomPermissionType } from './types/CustomPermissionType';
+import {
+  CustomPermission,
+  CustomPermissionName,
+  CustomPermissionStatus,
+  CustomPermissionResult,
+} from './types/CustomPermission';
 
 /** @internal */
 const miniAppInstance = new MiniApp();
@@ -18,7 +23,10 @@ export default miniAppInstance;
 export {
   AdTypes,
   InterstitialAdResponse,
+  CustomPermission,
+  CustomPermissionName,
+  CustomPermissionStatus,
+  CustomPermissionResult,
   RewardedAdResponse,
   ShareInfoType,
-  CustomPermissionType,
 };