+++ js-miniapp-sdk/src/miniapp.ts
@@ -1,5 +1,9 @@
 import { InterstitialAdResponse } from './types/responseTypes/interstitial';
-import { MiniAppPermissionType } from './types/MiniAppPermissionType';
+import {
+  CustomPermission,
+  CustomPermissionResult,
+} from './types/CustomPermission';
+import { DevicePermission } from './types/DevicePermission';
 import { RewardedAdResponse } from './types/responseTypes/rewarded';
 import { ShareInfoType } from './types/ShareInfoType';
 
@@ -12,6 +16,19 @@ interface MiniAppFeatures {
 
   /** @returns The Promise of permission result of mini app from injected side. */
   requestLocationPermission(): Promise<string>;
+
+  /**
+   *
+   * Request that the user grant custom permissions related to accessing user data.
+   * Typically, this will show a dialog in the Host App asking the user grant access to your Mini App.
+   * You can pass multiple permissions at once and the Host App will request all of those permissions within a single dialog.
+   *
+   * @param permissions An array containing CustomPermission objects - permission name and description
+   * @returns Promise with the custom permission results - "ALLOWED" or "DENIED" for each permission
+   */
+  requestCustomPermissions(
+    permissions: CustomPermission[]
+  ): Promise<CustomPermissionResult[]>;
 
   /**
    * @param info The shared data must match the property in [ShareInfoType].
@@ -71,7 +88,15 @@ export class MiniApp implements MiniAppFeatures, Ad {
   }
 
   requestLocationPermission(): Promise<string> {
-    return this.requestPermission(MiniAppPermissionType.LOCATION);
+    return this.requestPermission(DevicePermission.LOCATION);
+  }
+
+  requestCustomPermissions(
+    permissions: CustomPermission[]
+  ): Promise<CustomPermissionResult[]> {
+    return (window as any).MiniAppBridge.requestCustomPermissions(
+      permissions
+    ).then(permissionResult => permissionResult.permissions);
   }
 
   loadInterstitialAd(id: string): Promise<null | Error> {