+++ js-miniapp-sdk/src/types/CustomPermissionType.ts
@@ -1,4 +1,0 @@
-export interface CustomPermissionType {
-  name: string;
-  description: string;
-}