+++ js-miniapp-sdk/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### v1.2.0 (In progress)
 
-- Added support for requesting the load and display of Interstitial & Rewarded Ads by the host application [See here](README.MD#4-Show-Ads)
+- **Feature:** Added support for requesting the load and display of Interstitial & Rewarded Ads by the host application [See here](README.MD#4-Show-Ads)
+- **Feature:** Added `MiniApp.requestCustomPermissions` for requesting `USER_NAME`, `PROFILE_PHOTO`, and `CONTACT_LIST` permissions.
 
 ### v1.1.0 (2020-7-21)